+++ main.tf
@@ -12,7 +12,7 @@ locals {
 
   create_bucket_acl = (var.acl != null && var.acl != "null") || length(local.grants) > 0
 
-  attach_policy = var.attach_require_latest_tls_policy || var.attach_access_log_delivery_policy || var.attach_elb_log_delivery_policy || var.attach_lb_log_delivery_policy || var.attach_deny_insecure_transport_policy || var.attach_inventory_destination_policy || var.attach_deny_incorrect_encryption_headers || var.attach_deny_incorrect_kms_key_sse || var.attach_deny_unencrypted_object_uploads || var.attach_deny_ssec_encrypted_object_uploads || var.attach_policy || var.attach_waf_log_delivery_policy
+  attach_policy = var.attach_require_latest_tls_policy || var.attach_access_log_delivery_policy || var.attach_elb_log_delivery_policy || var.attach_lb_log_delivery_policy || var.attach_cloudtrail_log_delivery_policy || var.attach_deny_insecure_transport_policy || var.attach_inventory_destination_policy || var.attach_deny_incorrect_encryption_headers || var.attach_deny_incorrect_kms_key_sse || var.attach_deny_unencrypted_object_uploads || var.attach_deny_ssec_encrypted_object_uploads || var.attach_policy || var.attach_waf_log_delivery_policy
 
   # Variables with type `any` should be jsonencode()'d when value is coming from Terragrunt
   grants               = try(jsondecode(var.grant), var.grant)
@@ -25,6 +25,8 @@ locals {
 resource "aws_s3_bucket" "this" {
   count = local.create_bucket && !var.is_directory_bucket ? 1 : 0
 
+  region = var.region
+
   bucket        = var.bucket
   bucket_prefix = var.bucket_prefix
 
@@ -35,6 +37,8 @@ resource "aws_s3_bucket" "this" {
 
 resource "aws_s3_directory_bucket" "this" {
   count = local.create_bucket && var.is_directory_bucket ? 1 : 0
+
+  region = var.region
 
   bucket          = "${var.bucket}--${var.availability_zone_id}--x-s3"
   data_redundancy = var.data_redundancy
@@ -49,6 +53,8 @@ resource "aws_s3_directory_bucket" "this" {
 
 resource "aws_s3_bucket_logging" "this" {
   count = local.create_bucket && length(keys(var.logging)) > 0 && !var.is_directory_bucket ? 1 : 0
+
+  region = var.region
 
   bucket = aws_s3_bucket.this[0].id
 
@@ -78,6 +84,8 @@ resource "aws_s3_bucket_logging" "this" {
 
 resource "aws_s3_bucket_acl" "this" {
   count = local.create_bucket && local.create_bucket_acl && !var.is_directory_bucket ? 1 : 0
+
+  region = var.region
 
   bucket                = aws_s3_bucket.this[0].id
   expected_bucket_owner = var.expected_bucket_owner
@@ -117,6 +125,8 @@ resource "aws_s3_bucket_acl" "this" {
 
 resource "aws_s3_bucket_website_configuration" "this" {
   count = local.create_bucket && length(keys(var.website)) > 0 && !var.is_directory_bucket ? 1 : 0
+
+  region = var.region
 
   bucket                = aws_s3_bucket.this[0].id
   expected_bucket_owner = var.expected_bucket_owner
@@ -173,6 +183,8 @@ resource "aws_s3_bucket_website_configuration" "this" {
 resource "aws_s3_bucket_versioning" "this" {
   count = local.create_bucket && length(keys(var.versioning)) > 0 && !var.is_directory_bucket ? 1 : 0
 
+  region = var.region
+
   bucket                = aws_s3_bucket.this[0].id
   expected_bucket_owner = var.expected_bucket_owner
   mfa                   = try(var.versioning["mfa"], null)
@@ -188,6 +200,8 @@ resource "aws_s3_bucket_versioning" "this" {
 
 resource "aws_s3_bucket_server_side_encryption_configuration" "this" {
   count = local.create_bucket && length(keys(var.server_side_encryption_configuration)) > 0 ? 1 : 0
+
+  region = var.region
 
   bucket                = var.is_directory_bucket ? aws_s3_directory_bucket.this[0].bucket : aws_s3_bucket.this[0].id
   expected_bucket_owner = var.expected_bucket_owner
@@ -213,6 +227,8 @@ resource "aws_s3_bucket_server_side_encryption_configuration" "this" {
 resource "aws_s3_bucket_accelerate_configuration" "this" {
   count = local.create_bucket && var.acceleration_status != null && !var.is_directory_bucket ? 1 : 0
 
+  region = var.region
+
   bucket                = aws_s3_bucket.this[0].id
   expected_bucket_owner = var.expected_bucket_owner
 
@@ -223,6 +239,8 @@ resource "aws_s3_bucket_accelerate_configuration" "this" {
 resource "aws_s3_bucket_request_payment_configuration" "this" {
   count = local.create_bucket && var.request_payer != null && !var.is_directory_bucket ? 1 : 0
 
+  region = var.region
+
   bucket                = aws_s3_bucket.this[0].id
   expected_bucket_owner = var.expected_bucket_owner
 
@@ -232,6 +250,8 @@ resource "aws_s3_bucket_request_payment_configuration" "this" {
 
 resource "aws_s3_bucket_cors_configuration" "this" {
   count = local.create_bucket && length(local.cors_rules) > 0 && !var.is_directory_bucket ? 1 : 0
+
+  region = var.region
 
   bucket                = aws_s3_bucket.this[0].id
   expected_bucket_owner = var.expected_bucket_owner
@@ -252,6 +272,8 @@ resource "aws_s3_bucket_cors_configuration" "this" {
 
 resource "aws_s3_bucket_lifecycle_configuration" "this" {
   count = local.create_bucket && length(local.lifecycle_rules) > 0 ? 1 : 0
+
+  region = var.region
 
   bucket                                 = var.is_directory_bucket ? aws_s3_directory_bucket.this[0].bucket : aws_s3_bucket.this[0].id
   expected_bucket_owner                  = var.expected_bucket_owner
@@ -369,6 +391,8 @@ resource "aws_s3_bucket_lifecycle_configuration" "this" {
 resource "aws_s3_bucket_object_lock_configuration" "this" {
   count = local.create_bucket && var.object_lock_enabled && try(var.object_lock_configuration.rule.default_retention, null) != null ? 1 : 0
 
+  region = var.region
+
   bucket                = aws_s3_bucket.this[0].id
   expected_bucket_owner = var.expected_bucket_owner
   token                 = try(var.object_lock_configuration.token, null)
@@ -384,6 +408,8 @@ resource "aws_s3_bucket_object_lock_configuration" "this" {
 
 resource "aws_s3_bucket_replication_configuration" "this" {
   count = local.create_bucket && length(keys(var.replication_configuration)) > 0 && !var.is_directory_bucket ? 1 : 0
+
+  region = var.region
 
   bucket = aws_s3_bucket.this[0].id
   role   = var.replication_configuration["role"]
@@ -550,6 +576,8 @@ resource "aws_s3_bucket_replication_configuration" "this" {
 resource "aws_s3_bucket_policy" "this" {
   count = local.create_bucket && local.attach_policy ? 1 : 0
 
+  region = var.region
+
   # Chain resources (s3_bucket -> s3_bucket_public_access_block -> s3_bucket_policy )
   # to prevent "A conflicting conditional operation is currently in progress against this resource."
   # Ref: https://github.com/hashicorp/terraform-provider-aws/issues/7628
@@ -569,6 +597,7 @@ data "aws_iam_policy_document" "combined" {
     var.attach_elb_log_delivery_policy ? data.aws_iam_policy_document.elb_log_delivery[0].json : "",
     var.attach_lb_log_delivery_policy ? data.aws_iam_policy_document.lb_log_delivery[0].json : "",
     var.attach_access_log_delivery_policy ? data.aws_iam_policy_document.access_log_delivery[0].json : "",
+    var.attach_cloudtrail_log_delivery_policy ? data.aws_iam_policy_document.cloudtrail_log_delivery[0].json : "",
     var.attach_require_latest_tls_policy ? data.aws_iam_policy_document.require_latest_tls[0].json : "",
     var.attach_deny_insecure_transport_policy ? data.aws_iam_policy_document.deny_insecure_transport[0].json : "",
     var.attach_deny_unencrypted_object_uploads ? data.aws_iam_policy_document.deny_unencrypted_object_uploads[0].json : "",
@@ -619,7 +648,7 @@ data "aws_iam_policy_document" "elb_log_delivery" {
 
   # Policy for AWS Regions created before August 2022 (e.g. US East (N. Virginia), Asia Pacific (Singapore), Asia Pacific (Sydney), Asia Pacific (Tokyo), Europe (Ireland))
   dynamic "statement" {
-    for_each = { for k, v in local.elb_service_accounts : k => v if k == data.aws_region.current.name }
+    for_each = { for k, v in local.elb_service_accounts : k => v if k == data.aws_region.current.region }
 
     content {
       sid = format("ALBRegion%s", title(statement.key))
@@ -707,7 +736,7 @@ data "aws_iam_policy_document" "lb_log_delivery" {
   count = local.create_bucket && var.attach_lb_log_delivery_policy && !var.is_directory_bucket ? 1 : 0
 
   statement {
-    sid = "AWSLogDeliveryWrite"
+    sid = "AlbNlbLogDeliveryWrite"
 
     principals {
       type        = "Service"
@@ -762,7 +791,7 @@ data "aws_iam_policy_document" "lb_log_delivery" {
   }
 
   statement {
-    sid = "AWSLogDeliveryAclCheck"
+    sid = "AlbNlbLogDeliveryAclCheck"
 
     effect = "Allow"
 
@@ -902,7 +931,7 @@ data "aws_iam_policy_document" "waf_log_delivery" {
   count = local.create_bucket && var.attach_waf_log_delivery_policy && !var.is_directory_bucket ? 1 : 0
 
   statement {
-    sid = "AWSLogDeliveryWrite"
+    sid = "WafLogDeliveryWrite"
 
     effect = "Allow"
 
@@ -933,13 +962,13 @@ data "aws_iam_policy_document" "waf_log_delivery" {
 
     condition {
       test     = "ArnLike"
-      values   = ["arn:aws:logs:${data.aws_region.current.name}:${data.aws_caller_identity.current.id}:*"]
+      values   = ["arn:aws:logs:${data.aws_region.current.region}:${data.aws_caller_identity.current.id}:*"]
       variable = "aws:SourceArn"
     }
   }
 
   statement {
-    sid = "AWSLogDeliveryAclCheck"
+    sid = "WafLogDeliveryAclCheck"
 
     effect = "Allow"
 
@@ -964,8 +993,48 @@ data "aws_iam_policy_document" "waf_log_delivery" {
 
     condition {
       test     = "ArnLike"
-      values   = ["arn:aws:logs:${data.aws_region.current.name}:${data.aws_caller_identity.current.id}:*"]
+      values   = ["arn:aws:logs:${data.aws_region.current.region}:${data.aws_caller_identity.current.id}:*"]
       variable = "aws:SourceArn"
+    }
+  }
+}
+
+# CloudTrail
+data "aws_iam_policy_document" "cloudtrail_log_delivery" {
+  count = local.create_bucket && var.attach_cloudtrail_log_delivery_policy && !var.is_directory_bucket ? 1 : 0
+
+  statement {
+    sid = "AWSCloudTrailAclCheck"
+    principals {
+      type        = "Service"
+      identifiers = ["cloudtrail.amazonaws.com"]
+    }
+    actions = [
+      "s3:GetBucketAcl",
+    ]
+    resources = [
+      aws_s3_bucket.this[0].arn,
+    ]
+  }
+
+  statement {
+    sid = "AWSCloudTrailWrite"
+    principals {
+      type        = "Service"
+      identifiers = ["cloudtrail.amazonaws.com"]
+    }
+    actions = [
+      "s3:PutObject",
+    ]
+    resources = [
+      "${aws_s3_bucket.this[0].arn}/AWSLogs/*",
+    ]
+    condition {
+      test     = "StringEquals"
+      variable = "s3:x-amz-acl"
+      values = [
+        "bucket-owner-full-control",
+      ]
     }
   }
 }
@@ -1147,6 +1216,8 @@ data "aws_iam_policy_document" "deny_ssec_encrypted_object_uploads" {
 resource "aws_s3_bucket_public_access_block" "this" {
   count = local.create_bucket && var.attach_public_policy && !var.is_directory_bucket ? 1 : 0
 
+  region = var.region
+
   bucket = aws_s3_bucket.this[0].id
 
   block_public_acls       = var.block_public_acls
@@ -1157,6 +1228,8 @@ resource "aws_s3_bucket_public_access_block" "this" {
 
 resource "aws_s3_bucket_ownership_controls" "this" {
   count = local.create_bucket && var.control_object_ownership && !var.is_directory_bucket ? 1 : 0
+
+  region = var.region
 
   bucket = local.attach_policy ? aws_s3_bucket_policy.this[0].id : aws_s3_bucket.this[0].id
 
@@ -1174,6 +1247,8 @@ resource "aws_s3_bucket_ownership_controls" "this" {
 
 resource "aws_s3_bucket_intelligent_tiering_configuration" "this" {
   for_each = { for k, v in local.intelligent_tiering : k => v if local.create_bucket && !var.is_directory_bucket }
+
+  region = var.region
 
   name   = each.key
   bucket = aws_s3_bucket.this[0].id
@@ -1203,6 +1278,8 @@ resource "aws_s3_bucket_intelligent_tiering_configuration" "this" {
 resource "aws_s3_bucket_metric" "this" {
   for_each = { for k, v in local.metric_configuration : k => v if local.create_bucket && !var.is_directory_bucket }
 
+  region = var.region
+
   name   = each.value.name
   bucket = aws_s3_bucket.this[0].id
 
@@ -1217,6 +1294,8 @@ resource "aws_s3_bucket_metric" "this" {
 
 resource "aws_s3_bucket_inventory" "this" {
   for_each = { for k, v in var.inventory_configuration : k => v if local.create_bucket && !var.is_directory_bucket }
+
+  region = var.region
 
   name                     = each.key
   bucket                   = try(each.value.bucket, aws_s3_bucket.this[0].id)
@@ -1318,6 +1397,8 @@ data "aws_iam_policy_document" "inventory_and_analytics_destination_policy" {
 
 resource "aws_s3_bucket_analytics_configuration" "this" {
   for_each = { for k, v in var.analytics_configuration : k => v if local.create_bucket && !var.is_directory_bucket }
+
+  region = var.region
 
   bucket = aws_s3_bucket.this[0].id
   name   = each.key